migrated backup_file_s3 method inside s3_updnload utility.
